+++ NEWS.md
@@ -3,3 +3,4 @@
 * Added NEWS file
 * Added verbosity to `list_surveys()`
 * Added `get_citation()` - Resolves #38
+* Added bibtex citation style as default to `get_citation()` - Resolves #52